src/credit_scoring_kedro/settings.py
"""Project settings. There is no need to edit this file unless you want to change values
from the Kedro defaults. For further information, including these default values, see
https://docs.kedro.org/en/stable/kedro_project_setup/settings.html."""

# src/credit_scoring_kedro/settings.py
# Instantiated project hooks.
# Hooks are executed in a Last-In-First-Out (LIFO) order.
# Ordre d'exécution (LIFO) : MLflowSetupHook → ArtifactsExportHook → LoadEnvVarsHook
from pathlib import Path

from dotenv import load_dotenv

from credit_scoring_kedro.hooks import (
    ArtifactsExportHook,
    MLflowSetupHook,
)

# Charger le .env AVANT kedro-mlflow (qui lit mlflow.yml dans after_context_created)
load_dotenv(Path(__file__).parent.parent.parent / ".env", override=True)


HOOKS = (
    ArtifactsExportHook(),  # 2. Exporte les artefacts vers data/api/ après kedro run
    MLflowSetupHook(),  # 3. Configure les tags MLflow avant chaque pipeline
)

# Installed plugins for which to disable hook auto-registration.
# DISABLE_HOOKS_FOR_PLUGINS = ("kedro-viz",)

# Class that manages storing KedroSession data.
# from kedro.framework.session.store import BaseSessionStore
# SESSION_STORE_CLASS = BaseSessionStore
# Keyword arguments to pass to the `SESSION_STORE_CLASS` constructor.
# SESSION_STORE_ARGS = {
#     "path": "./sessions"
# }

# Directory that holds configuration.
# CONF_SOURCE = "conf"

# Class that manages how configuration is loaded.
# from kedro.config import OmegaConfigLoader

# CONFIG_LOADER_CLASS = OmegaConfigLoader

# Keyword arguments to pass to the `CONFIG_LOADER_CLASS` constructor.
CONFIG_LOADER_ARGS = {
    "base_env": "base",
    "default_run_env": "local",
    # "config_patterns": {
    #     "spark" : ["spark*/"],
    #     "parameters": ["parameters*", "parameters*/**", "**/parameters*"],
    # }
}

# Class that manages Kedro's library components.
# from kedro.framework.context import KedroContext
# CONTEXT_CLASS = KedroContext

# Class that manages the Data Catalog.
# from kedro.io import DataCatalog
# DATA_CATALOG_CLASS = DataCatalog
