pyproject.toml
[build-system]
requires = ["setuptools>=68,<72"]
build-backend = "setuptools.build_meta"

[project]
name = "credit_scoring_kedro"
version = "1.0.0"
description = "Projet MLOps de scoring de crédit avec Kedro — Formation MLOps ORSYS"
requires-python = ">=3.12"
readme = "README.md"
dependencies = [
    # ── Kedro core ────────────────────────────────────────────────────────────
    "kedro~=0.19.14",
    "kedro-mlflow==0.12.2",
    "kedro-datasets[pandas.CSVDataset,matplotlib.MatplotlibWriter,pickle.PickleDataset,json.JSONDataset]>=3.0.0",
    # ── ML ────────────────────────────────────────────────────────────────────
    "scikit-learn>=1.5.0",
    "pandas>=2.2.0",
    "numpy>=1.26.0",
    "matplotlib>=3.8.0",
    # mlflow >= 2.17 ne dépend plus de pkg_resources (Python 3.12 safe)
    "mlflow>=2.17.0",
    "joblib>=1.4.0",
    # ── API FastAPI ───────────────────────────────────────────────────────────
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "pydantic>=2.9.0",
    "python-multipart>=0.0.12",
    # ── Sécurité JWT + Rate Limiting ──────────────────────────────────────────
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "bcrypt==4.0.1",          # épinglé : passlib 1.7.4 incompatible avec bcrypt>=5
    "slowapi>=0.1.9",
    "python-dotenv>=1.0.0",
    # setuptools < 72 fournit pkg_resources (shim pour les libs qui en dépendent encore)
    # mlflow >= 2.17 n'en a plus besoin, mais on garde pour kedro-mlflow
    "setuptools>=68,<72",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=5.0.0",
    "httpx>=0.27.0",
    "ruff>=0.6.0",
    "ipykernel>=6.0.0",
    "kedro-viz>=9.0.0",
]
load-testing = [
    "locust>=2.31.0",
]
# ─── Cloud extras ────────────────────────────────────────────────────────────
# NOTE : installer UN SEUL extra cloud à la fois (conflits fsspec entre s3fs et adlfs).
# Usage :
#   uv sync --extra cloud-aws    # pour AWS S3
#   uv sync --extra cloud-gcp    # pour GCP GCS
#   uv sync --extra cloud-azure  # pour Azure Blob
cloud-aws   = ["s3fs>=2024.6.0"]
cloud-gcp   = ["gcsfs>=2024.6.0"]
cloud-azure = ["adlfs>=2024.7.0"]

[project.scripts]
credit-scoring-kedro = "credit_scoring_kedro.__main__:main"
credit-scoring-api   = "credit_scoring_kedro.api.main:main"

# ─── uv ──────────────────────────────────────────────────────────────────────
[tool.uv]
# Résolution stricte : garantit la reproductibilité du uv.lock
resolution = "lowest-direct"

[tool.uv.sources]
# Pas de sources custom — tout vient de PyPI

# ─── Kedro ───────────────────────────────────────────────────────────────────
[tool.kedro]
package_name = "credit_scoring_kedro"
project_name = "credit-scoring-kedro"
kedro_init_version = "0.19.14"
tools = "['None']"
example_pipeline = "False"
source_dir = "src"

[project.entry-points."kedro.hooks"]

[tool.setuptools.dynamic]

[tool.setuptools.packages.find]
where = ["src"]
namespaces = false

# ─── Ruff ────────────────────────────────────────────────────────────────────
[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]
ignore = ["E501"]

# ─── Pytest ──────────────────────────────────────────────────────────────────
[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --tb=short"

[tool.coverage.run]
omit = [
    "src/credit_scoring_kedro/pipelines/*",
    "src/credit_scoring_kedro/hooks.py",
    "src/credit_scoring_kedro/pipeline_registry.py",
    "src/credit_scoring_kedro/__main__.py",
    "src/credit_scoring_kedro/settings.py",
]


[tool.kedro_telemetry]
project_id = "043731c72df74cc2bc71817647df664d"

[dependency-groups]
dev = [
    "pre-commit>=3.6.0"
]
